+++ tomcat/src/main/java/org/apache/coyote/Controller.java
@@ -1,7 +1,7 @@
 package org.apache.coyote;
 
-import org.apache.coyote.http11.request.HttpRequest;
-import org.apache.coyote.http11.response.HttpResponse;
+import org.apache.coyote.http11.message.request.HttpRequest;
+import org.apache.coyote.http11.message.response.HttpResponse;
 
 public interface Controller {
