+++ tomcat/src/main/java/org/apache/coyote/HomeController.java
@@ -1,8 +1,8 @@
 package org.apache.coyote;
 
-import org.apache.coyote.http11.request.HttpRequest;
-import org.apache.coyote.http11.response.HttpResponse;
-import org.apache.coyote.http11.response.HttpStatus;
+import org.apache.coyote.http11.message.request.HttpRequest;
+import org.apache.coyote.http11.message.response.HttpResponse;
+import org.apache.coyote.http11.message.response.HttpStatus;
 
 public class HomeController extends AbstractController {
