+++ tomcat/src/main/java/org/apache/coyote/StaticResourceController.java
@@ -3,8 +3,8 @@ package org.apache.coyote;
 import java.net.URL;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import org.apache.coyote.http11.request.HttpRequest;
-import org.apache.coyote.http11.response.HttpResponse;
+import org.apache.coyote.http11.message.request.HttpRequest;
+import org.apache.coyote.http11.message.response.HttpResponse;
 
 public class StaticResourceController extends AbstractController {
