+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Processor.java
@@ -8,13 +8,11 @@ import java.io.InputStreamReader;
 import java.io.OutputStream;
 import java.net.Socket;
 import java.net.URISyntaxException;
-import org.apache.commons.lang3.StringUtils;
 import org.apache.coyote.AbstractController;
 import org.apache.coyote.Processor;
 import org.apache.coyote.RequestMapping;
-import org.apache.coyote.http11.request.HttpRequest;
-import org.apache.coyote.http11.request.HttpRequestLine;
-import org.apache.coyote.http11.response.HttpResponse;
+import org.apache.coyote.http11.message.request.HttpRequest;
+import org.apache.coyote.http11.message.response.HttpResponse;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
