+++ tomcat/src/main/java/org/apache/coyote/http11/message/common/HttpBody.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.message.common;
 
 public class HttpBody {
