+++ tomcat/src/main/java/org/apache/coyote/http11/message/common/HttpHeaders.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.message.common;
 
 import java.util.HashMap;
 import java.util.Map;
